requirements.txt
sentence-transformers
faiss-cpu
transformers
torch
